requirements.txt
flask
openai
pytest
python-dotenv
pyngrok
GitPython